+++ tests/test/callback.js
@@ -3,11 +3,13 @@ var atoum = require('../..')(module),
     testedClass = require('../../lib/test/callback'),
     unit = module.exports = {
         testClass: function() {
-            var object;
+            var object, otherObject;
 
             this
                 .function(object = testedClass())
                 .object(object.controller).isInstanceOf(controller)
+                .function(otherObject = new testedClass()).isNotIdenticalTo(object)
+                .object(otherObject.controller).isInstanceOf(controller)
             ;
         }
     };